Add files via upload
This is just a basic player movement script for Unity 3d, it opperates as a FPS controller using CharacterController and no RigidBody. Ground is the tag I use for ground detection so make sure you have that set properly. Let me know if there's any questions or improvements I could add, I plan on updating it with whatever change's I end up adding to it but for now it's good for general use.